ref: Use strings for data routes (#23)
* Convert data route to accept Strings instead of Enum values

* Update CHANGELOG

* Update README
